Implement unreachable code elimination via "call graph visit"
in `remove_unreachable_code`, the algorithm tries
to visit every instruction, tagging visited "nodes/instructions" as
such.
when every possible path is exhausted (= there are no more paths
to be taken), the algorithm generates a new list of instructions,
ordering them by position in the original "ast" and returns it.