Fix multiple labels with same name at the same instruction
If an instruction is reachable from N locations in the code
via jump/condjump, there would be N labels associated to the
instructions.
Fix it to remove multiple instances of the same label and replace
them with just one instance.
When using the `compress_jumps` pass before `remove_unreachable`,
we should always have only one label associated to each instruction.
Labels are sorted: this detail does not change semantics, as the
order of labels on an instruction does not change the execution
of the program.