Handle condjumps with no instructions, mark them as "_hrm_unreachable"
In some example programs, conditional jumps may try to reach
a label that is associated to no instructions.
This kind of conditional jumps are now replaced with a conditional
jump to a special label named `_hrm_unreachable`, to make it clear
that, if the condition is true, the computation is terminated.
Fix the handling of this condition in the `remove_unreachable` pass,
to avoid trying to jump at non-existent instructions.

With this patch, the following program
```
emp = inbox
if nz then
  outbox
endif
```
translates as

```
inbox
jez _hrm_unreachable
outbox
```