Fix operations missing from compiled code in non-looping programs
When the `remove_unreachable` pass was written, it was _implicitely_
assumed that the code would loop and we wouldn't have any
path we did not explore.
Programs that users write for the game actually _are_ looping
- so we would have an opportunity to reach an already visited
instruction -, but they may not be looping.

This commit removes the assumption and fixed `remove_unreachable`
to work correctly with non-looping inputs. The change should
have also fixed possible (but untested) cases where missed jcond
paths are unexplored in looping programs.